refactor: add re-exports to pages and components
