+++ src/components/index.ts
@@ -1,0 +1,9 @@
+export { default as Footer } from './Footer';
+
+export { default as Header } from './Header';
+
+export { default as Loader } from './Loader';
+
+export { default as Modal } from './Modal';
+
+export { default as Toast } from './Toast';
