+++ src/pages/TemplatePage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './templatePage';
