+++ src/pages/index.ts
@@ -1,0 +1,15 @@
+export { default as BoardsPage } from './BoardsPage';
+
+export { default as MainPage } from './MainPage';
+
+export { default as NotFoundPage } from './NotFoundPage';
+
+export { default as SignInPage } from './SignInPage';
+
+export { default as SignUpPage } from './SignUpPage';
+
+export { default as SingleBoardPage } from './SingleBoardPage';
+
+export { default as TemplatePage } from './TemplatePage';
+
+export { default as UserPage } from './UserPage';
